Refactor Protobuf utility class to webapp
* It's not used in library
