+++ gtfs-realtime-validator-webapp/src/main/java/edu/usf/cutr/gtfsrtvalidator/servlets/GetFeedJSON.java
@@ -17,7 +17,7 @@
 
 package edu.usf.cutr.gtfsrtvalidator.servlets;
 
-import edu.usf.cutr.gtfsrtvalidator.lib.helper.ProtoBuf;
+import edu.usf.cutr.gtfsrtvalidator.util.ProtoBufUtil;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -31,7 +31,7 @@ public class GetFeedJSON extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response){
         String value = request.getParameter("path");
-        String jsonString = ProtoBuf.protoToJSON(value);
+        String jsonString = ProtoBufUtil.protoToJSON(value);
 
         response.setContentType("application/json");
         response.setStatus(HttpServletResponse.SC_OK);
